pyproject.toml
[tool.poetry]
name = "langgraph-tutorials"
version = "0.1.0"
description = ""
authors = ["Yohei <y.meguro.2008@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
python-dotenv = "^1.0.1"
ruff = "^0.7.3"
langgraph = "^0.2.45"

[tool.poetry.group.quick_start.dependencies]
langsmith = "^0.1.142"
langchain-anthropic = "^0.2.4"
ipython = "^8.29.0"
langchain-openai = "^0.2.6"
tavily-python = "^0.5.0"
langchain-community = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
